src/requirements.txt
PyYAML==6.0
pandas==1.4.2
imblearn==0.0
spacy==3.2.1
scikit-learn==1.0.2
bayesian-optimization==1.2.0
gensim==4.1.2